Cybersecurity Automation Platform (CAP)/requirements.txt
flask
flask-cors
requests
scapy
pandas
numpy
scikit-learn
joblib
pymongo
python-nmap
dnspython
python-dotenv
